+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- Add `sticky` prop on `EmojiPicker.Root` to allow disabling sticky category headers, thanks @Earthsplit!
 - Fix return types to support TypeScript versions lower than 5.1.
 
 ## [0.2.0] - 2025-04-02